Add last comments on code

+++ src/main/java/com/mytaxi/service/car/DefaultCarService.java
@@ -3,8 +3,7 @@ package com.mytaxi.service.car;
 import com.mytaxi.dataaccessobject.CarRepository;
 import com.mytaxi.dataaccessobject.specifications.CarSpecifications;
 import com.mytaxi.domainobject.CarDO;
-import com.mytaxi.domainobject.DriverDO;
-import com.mytaxi.domainobject.ManufacturerDO;
+import com.mytaxi.exception.BusinessRuleException;
 import com.mytaxi.exception.ConstraintsViolationException;
 import com.mytaxi.exception.EntityNotFoundException;
 import com.mytaxi.service.AbstractService;
@@ -16,8 +15,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
+/**
+ * This class handles all business rules on {@link CarDO} entities
+ */
 @Service
 public class DefaultCarService extends AbstractService<CarDO, Long> implements CarService {
 
@@ -31,30 +32,64 @@ public class DefaultCarService extends AbstractService<CarDO, Long> implements C
     }
 
 
+    /**
+     * Find all cars.
+     * @return The list of {@link CarDO}
+     * @throws EntityNotFoundException if no cars found
+     */
     @Override
-    public List<CarDO> find() {
+    public List<CarDO> find() throws EntityNotFoundException {
         return super.findAll();
     }
 
+    /**
+     * Find a car by id
+     * @param id The car's id
+     * @return The {@link CarDO} fpund
+     * @throws EntityNotFoundException if no car found
+     */
     @Override
     @Transactional(readOnly = true)
     public CarDO find(Long id) throws EntityNotFoundException {
         return super.findById(id);
     }
 
+    /**
+     * Find a car by its assigned driver
+     * @param driverId The driver's id
+     * @return The {@link CarDO} found
+     * @throws EntityNotFoundException if no car is assigned to user
+     */
     @Override
     @Transactional(readOnly = true)
     public CarDO findByDriverId(long driverId) throws EntityNotFoundException {
         return super.findOne(CarSpecifications.findByAssignedDriver(driverId));
     }
 
+    /**
+     * Create a car
+     * @param carDO The {@link CarDO} provided
+     * @return The created {@link CarDO} with generated id.
+     * @throws ConstraintsViolationException if some db constraints were violated
+     */
     @Override
     public CarDO create(CarDO carDO) throws ConstraintsViolationException {
         return super.save(carDO);
     }
 
+    /**
+     * Delete a car by id
+     * @param id The car's id
+     * @throws EntityNotFoundException if the car doen't exist
+     * @throws BusinessRuleException if the car is assigned currently to a driver
+     */
     @Override
-    public void delete(Long id) throws EntityNotFoundException {
+    public void delete(Long id) throws EntityNotFoundException, BusinessRuleException {
+        CarDO carDO = super.findById(id);
+        //validate if is not assigned to a driver
+        if (carDO.getDriverDO() != null) {
+            throw new BusinessRuleException("The car is currently assigned to a driver.");
+        }
         super.deleteById(id);
     }
